arm: tegra12: clocks: Update readl/writel calls
readl and writel now take void * instead of u32. Update the way they
are called to reflect this change.

Reviewed-on: http://git-master/r/196096

+++ arch/arm/mach-tegra/tegra12_clocks.c
@@ -6867,7 +6867,7 @@ bool tegra_clk_is_parent_allowed(struct clk *c, struct clk *p)
 void __init tegra12x_clk_init_la(void)
 {
 	struct clk *c;
-	u32 reg = readl((u32)misc_gp_base + MISC_GP_TRANSACTOR_SCRATCH_0);
+	u32 reg = readl(misc_gp_base + MISC_GP_TRANSACTOR_SCRATCH_0);
 
 	if (!(reg & MISC_GP_TRANSACTOR_SCRATCH_LA_ENABLE))
 		return;
